requirements.txt
PyYAML>=5.4
PyJWT>=2.0.0
cryptography>=3.4.0